semana09/dia3/backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cloudinary==1.28.0
Django==3.2.9
djangorestframework==3.12.4
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
